backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
pillow==10.4.0
numpy==2.1.3
tflite-runtime==2.14.0
# If tflite-runtime is unavailable for your platform, use tensorflow==2.17.0 instead (heavier).
python-multipart==0.0.9
